refactor: replace _test.go suffix with //go:build test tag for cross-package test helpers
Test helper files that need to be accessible from other packages' tests
cannot use the _test.go suffix. Rename them back to .go files and add
//go:build test tags to keep them out of production builds.

Also consolidate verification test helpers into test_helpers.go by
merging test_helpers_test.go content.
