+++ internal/common/test_helpers.go
@@ -1,4 +1,6 @@
-//
+//go:build test
+// +build test
+
 //nolint:revive // var-naming: package name "common" is intentional for shared internal utilities
 package common
