+++ internal/runner/resource/test_helpers.go
@@ -1,3 +1,6 @@
+//go:build test
+// +build test
+
 package resource
 
 import (
