+++ internal/runner/security/test_helpers.go
@@ -1,3 +1,6 @@
+//go:build test
+// +build test
+
 package security
 
 import (
